Use consistent return type for imc_sendmsg()/imc_recvmsg() in test
Use "int" as the return type instead of "ssize_t" in order to match
src/trusted/service_runtime/include/sys/nacl_syscalls.h.

This is currently necessary for this test to compile successfully
against x86-64 nacl-glibc.  The compile failure exposes a bug in
nacl-glibc (ssize_t differs between x86-32 and x86-64), which should
be fixed later, but it is good to keep the imc_*() functions
consistent too.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=1246
TEST=build with "platform=x86-64 --nacl_glibc"

Review URL: http://codereview.chromium.org/6073005

git-svn-id: svn://svn.chromium.org/native_client/trunk/src/native_client@4066 fcba33aa-ac0c-11dd-b9e7-8d5594d729c2

Former-commit-id: 4b4565e328877927a5a2e9fd9f9ad74a297ee528

+++ tests/srpc_message/srpc_message.c
@@ -380,7 +380,7 @@ int main(int argc, char* argv[]) {
 /*
  * Mocked versions of imc_sendmsg and imc_recvmsg using a queue of MessageList.
  */
-ssize_t imc_sendmsg(int desc, struct NaClImcMsgHdr const* header, int flags) {
+int imc_sendmsg(int desc, struct NaClImcMsgHdr const* header, int flags) {
   size_t i;
   size_t offset = 0;
   struct MessageList* elt;
@@ -417,7 +417,7 @@ ssize_t imc_sendmsg(int desc, struct NaClImcMsgHdr const* header, int flags) {
   return offset;
 }
 
-ssize_t imc_recvmsg(int desc, struct NaClImcMsgHdr* header, int flags) {
+int imc_recvmsg(int desc, struct NaClImcMsgHdr* header, int flags) {
   size_t i;
   size_t offset = 0;
   struct MessageList* elt;
